+++ Aula05/App.js
@@ -1,7 +1,8 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import HomeScreen from './screens/HomeScreen';
-import ProfileScreen from './screens/ProfileScreen'
+import Perfil from './screens/ProfileScreen';
+import Galeria from './screens/Galeria';
 
 const Stack = createNativeStackNavigator();
 
@@ -10,8 +11,9 @@ export default function App() {
     <NavigationContainer>
       <Stack.Navigator>
         <Stack.Screen name="Início" component={HomeScreen} />
-        <Stack.Screen name="Perfil" component={ProfileScreen} />
+        <Stack.Screen name="Perfil" component={Perfil}/>
+        <Stack.Screen name="Galeria" component={Galeria}/>
       </Stack.Navigator>
     </NavigationContainer>
   );
-};
+}
